requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
crispy-bootstrap5==2024.2
Django==4.2
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.36
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1